Add comments, move Files to elevatorSimulation package to avoid default package
